requirements.txt
streamlit
tensorflow-cpu
numpy<2.0.0
pillow
joblib
scikit-learn
